+++ Integration/Helper.php
@@ -9,6 +9,8 @@
 
 namespace Prism\Integration;
 
+use Joomla\Utilities\ArrayHelper;
+
 defined('JPATH_PLATFORM') or die;
 
 /**
@@ -20,6 +22,15 @@ defined('JPATH_PLATFORM') or die;
  */
 abstract class Helper
 {
+    protected static $lookup;
+
+    /**
+     * Cache of menu item IDs.
+     *
+     * @var array
+     */
+    protected static $menuItemIds = array();
+
     /**
      * Get a user ID from request.
      * This method should be used for pages,
@@ -62,5 +73,104 @@ abstract class Helper
         }
 
         return (int)$userId;
+    }
+
+    /**
+     * Get Itemid of community builder that points to user profile.
+     *
+     * <code>
+     * $userId = Prism\Integration\Helper::getUserId();
+     * </code>
+     *
+     * @param string $componentName
+     * @param array $needles
+     *
+     * @return int
+     */
+    public static function getItemId($componentName, $needles)
+    {
+        // Prepare cache hash.
+        $hashString = "";
+        foreach ($needles as $view => $ids) {
+            // Break the loop and exit from the method,
+            // if it is not array with IDs.
+            if (!is_array($ids)) {
+                return null;
+            }
+
+            $ids = ArrayHelper::toInteger($ids);
+            $hashString .= $view . implode(",", $ids);
+        }
+        $hash = md5($hashString);
+
+        // Check the cache.
+        if (isset(self::$menuItemIds[$hash])) {
+            return self::$menuItemIds[$hash];
+        }
+
+        $app   = \JFactory::getApplication();
+        $menus = $app->getMenu('site');
+
+        // Prepare the reverse lookup array.
+        // Collect all menu items and create an array that contains
+        // the ID from the query string of the menu item as a key,
+        // and the menu item id (Itemid) as a value
+        // Example:
+        // array( "category" =>
+        //     1(id) => 100 (Itemid),
+        //     2(id) => 101 (Itemid)
+        // );
+        if (self::$lookup === null) {
+            self::$lookup = array();
+
+            $component = \JComponentHelper::getComponent($componentName);
+            $items     = $menus->getItems('component_id', $component->id);
+
+            if ($items) {
+                foreach ($items as $item) {
+                    if (isset($item->query) && isset($item->query['view'])) {
+                        $view = $item->query['view'];
+
+                        if (!isset(self::$lookup[$view])) {
+                            self::$lookup[$view] = array();
+                        }
+
+                        if (isset($item->query['id'])) {
+                            self::$lookup[$view][$item->query['id']] = $item->id;
+                        } else { // If it is a root element that have no a request parameter ID ( categories, authors ), we set 0 for an key
+                            self::$lookup[$view][0] = $item->id;
+                        }
+                    }
+                }
+            }
+        }
+
+        $result = null;
+        if ($needles) {
+
+            foreach ($needles as $view => $ids) {
+                if (isset(self::$lookup[$view])) {
+
+                    foreach ($ids as $id) {
+                        if (isset(self::$lookup[$view][(int)$id])) {
+                            $result = self::$lookup[$view][(int)$id];
+                            break;
+                        }
+                    }
+
+                }
+            }
+
+        } else {
+            $active = $menus->getActive();
+            if ($active) {
+                $result = $active->id;
+            }
+        }
+
+        // Set the menu item ID to the cache.
+        self::$menuItemIds[$hash] = $result;
+
+        return $result;
     }
 }