+++ Integration/Profiles/EasySocial.php
@@ -143,18 +143,16 @@ class EasySocial implements ProfilesInterface
      * $link = $profiles->getLink($userId);
      * </code>
      *
-     * @param integer $userId
+     * @param int $userId
+     * @param bool $route Route or not the link.
      *
      * @return string
      */
-    public function getLink($userId)
+    public function getLink($userId, $route = true)
     {
-        if (!isset($this->profiles[$userId])) {
-            $link = "";
-        } else {
-
+        $link = "";
+        if (isset($this->profiles[$userId]) and $route) {
             $options = array('id' => $this->getAlias($this->profiles[$userId]));
-
             $link = \FRoute::profile($options);
         }