+++ Integration/Profiles/ProfilesInterface.php
@@ -21,7 +21,7 @@ defined('JPATH_PLATFORM') or die;
 interface ProfilesInterface
 {
     public function load(array $ids);
-    public function getLink($userId);
+    public function getLink($userId, $route = true);
     public function getAvatar($userId, $size);
     public function getLocation($userId);
     public function getCountryCode($userId);